feat(trackeditor): flip from side-by-side to horizontal panes
Adds View menu item to flip the GUI arrangement from vertical to horizontal and
vice versa.

WARNING: Note that there's obviously issues with application state here. I
added some crude tracking of the currently-open track file, but there will be a
lot more to keep track of in the future, which will need a more robust way of
holding and restoring. Leaving that for now.